% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{read_raw_xlsx_data}
\alias{read_raw_xlsx_data}
\title{reading raw .xlsx data into R}
\usage{
read_raw_xlsx_data(file_name)
}
\arguments{
\item{file_name}{name of the .xlsx file}
}
\value{
This function returns an R object which contains the raw dataset.
}
\description{
This functions reads in raw .xlsx data into R (after performing a few
sanity checks) and throws an error if for any reason data cannot be read.
Note that this function takes in a file name, not file path. A relative file
path is automatically constructed in the function.
}
\examples{
read_raw_xlsx_data(file_name = "couponTransactions.xlsx")

}
